[tests] move project tests to algorithms.projection
